Known events which happened on October 1st before 9am

The Water Year in the UK runs from 9am of the 1st October of a given year to 8:59am of the 1st October of the next year. Since the WINFAP files contain information only on the date of the annual maximum (and not time) it is possible that an event is mis-classified when using the water_year function. This dataset lists the events which are known to have happened to October 1st before 9am. This is used to correct the WaterYear information in these known cases in the read_amax and get_amax functions. For some stations events on October 1st have been deemed as annual maxima only in some winfap releases. They are maintained in the dataset in the event that somebody read old winfap files.

known_Oct1

Format

A data frame with 36 rows and 3 variables:

Station

NRFA station number

Date

date of maximum flow (always the 1st October)

WaterYear

the correct water year for the peak flow

Source

Derived manually by identifying events which happened on Oct. 1st and comparing it with information on https://nrfa.ceh.ac.uk