A function to obtain annual maxima (AMAX) data using the NRFA API

The function queries the NRFA API for the .AM file similar to the WINFAP file for a given stations. It then processes the file in a fashion similar to read_amax.

get_amax(station)

Arguments

station

the NRFA station number for which the annual maxima records should be obtained. Can also be a vector of station numbers.

Value

a data.frame with information on the annual maxima for the station with the following columns

Station

NRFA station number (can be a vector of station numbers)

WaterYear

the correct water year for the peak flow

Date

date of maximum flow

Flow

the maximum flow in m3/s

Stage

the stage (height) reached by the river - this information is used to derive the flow via a rating curve

Rejected

logical, if TRUE the water year has been flagged as rejected by the NRFA

See also

Examples

  a40003 <- get_amax(40003) # the Medway at Teston / East Farleigh
  multipleStations <- get_amax(c(40003, 42003))
  names(multipleStations)
#> [1] "40003" "42003"
  summary(multipleStations$`42003`)
#>     Station        WaterYear         Date                          Flow      
#>  Min.   :42003   Min.   :1975   Min.   :1976-09-30 21:00:00   Min.   : 4.16  
#>  1st Qu.:42003   1st Qu.:1987   1st Qu.:1987-11-12 05:00:00   1st Qu.:15.71  
#>  Median :42003   Median :1999   Median :1999-12-25 02:15:00   Median :21.00  
#>  Mean   :42003   Mean   :1999   Mean   :2000-01-16 14:34:17   Mean   :25.05  
#>  3rd Qu.:42003   3rd Qu.:2011   3rd Qu.:2012-06-12 06:45:00   3rd Qu.:29.40  
#>  Max.   :42003   Max.   :2023   Max.   :2023-11-02 10:45:00   Max.   :81.18  
#>      Stage            Rejected      
#>  Min.   :-9999.000   Mode :logical  
#>  1st Qu.:    1.241   FALSE:29       
#>  Median :    1.383   TRUE :20       
#>  Mean   :-1427.148                  
#>  3rd Qu.:    1.621                  
#>  Max.   :    2.134